+++ R/utils.R
@@ -24,14 +24,17 @@ vcapply <- function(..., FUN.VALUE = character(1L)) {  # nolint object_name_lint
   vapply(..., FUN.VALUE = FUN.VALUE)
 }
 
+#' @keywords internal
 as_env <- function(x) {
   UseMethod("as_env")
 }
 
+#' @keywords internal
 as_env.character <- function(x) {
   getNamespace(x)
 }
 
+#' @keywords internal
 as_env.environment <- function(x) {
   x
 }
@@ -44,10 +47,12 @@ list_is_all_unnamed <- function(x) {
   is.null(names(x)) || all(names(x) == "")
 }
 
+#' @keywords internal
 raise <- function(x, ...) {
   UseMethod("raise")
 }
 
+#' @keywords internal
 raise.character <- function(x, ...) {
   x <- switch(x,
     "print" = , "info" = , "message" = message,
@@ -58,6 +63,7 @@ raise.character <- function(x, ...) {
   raise.function(x, ...)
 }
 
+#' @keywords internal
 raise.function <- function(x, msg, ...) {
   args <- list(msg, ...)